refactor: use ml-xy-convert for parsing
