+++ package.json
@@ -53,6 +53,7 @@
     "rollup": "^0.45.2"
   },
   "dependencies": {
+    "ml-xy-convert": "^1.0.0",
     "xy-parser": "^1.5.0"
   }
 }
